% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexiblas.R
\docType{package}
\name{flexiblas-package}
\alias{flexiblas-package}
\title{\pkg{flexiblas}: FlexiBLAS API Interface for \R}
\description{
FlexiBLAS is a BLAS wrapper library which allows to change the BLAS
without recompiling the programs.
}
\references{
Koehler M., Saak J. (2020).
"FlexiBLAS - A BLAS and LAPACK wrapper library with runtime exchangeable backends."
\doi{10.5281/zenodo.3909214}.
}
\author{
Martin Koehler, Iñaki Ucar
}
