% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{flexiblas_avail}
\alias{flexiblas_avail}
\title{Check Availability}
\usage{
flexiblas_avail()
}
\value{
A boolean.
}
\description{
Check whether FlexiBLAS is available.
}
\seealso{
\link{flexiblas_version}, \link{flexiblas-backends}, \link{flexiblas-threads}
}
