% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{flexiblas_version}
\alias{flexiblas_version}
\title{Get Version}
\usage{
flexiblas_version()
}
\value{
A \link{package_version} object.
}
\description{
Get current version of FlexiBLAS.
}
\seealso{
\link{flexiblas_avail}, \link{flexiblas-backends}, \link{flexiblas-threads}
}
