% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexiblas.R
\docType{package}
\name{flexiblas-package}
\alias{flexiblas-package}
\title{\pkg{flexiblas}: FlexiBLAS API Interface for \R}
\description{
Provides functions to switch the BLAS/LAPACK optimized backend
and change the number of threads without leaving the R session, which needs
to be linked against the FlexiBLAS wrapper library.
}
\references{
Koehler M., Saak J. (2020).
"FlexiBLAS - A BLAS and LAPACK wrapper library with runtime exchangeable backends."
\doi{10.5281/zenodo.3909214}.
}
\author{
Martin Koehler, Iñaki Ucar
}
